+++ R/drop_colr.R
@@ -3,7 +3,7 @@
 #' @param pattern A regular expression as in \code{grepl}.
 #' @param ... Arguments passed to \code{grepl}.
 #' @param checkDT If \code{TRUE} (the default), will error if \code{DT} is not a \code{data.table}.
-#' @export
+#' @export drop_colr drop_grep
 
 drop_colr <- function(DT, pattern, ..., checkDT = TRUE) {
   if (checkDT) {
@@ -12,4 +12,6 @@ drop_colr <- function(DT, pattern, ..., checkDT = TRUE) {
   
   DT[, .SD, .SDcols = names(DT)[!grepl(pattern = pattern, x = names(DT), ...)]]
 }
+
+drop_grep <- drop_colr