+++ NEWS.md
@@ -4,6 +4,9 @@
   - `isAttached` for conveniently determining whether a namespace is attached
   - `ahull` for locating rectangles in a plot, as for automatically locating a text box.
 
+* Minor changes:
+  - `if_else` reports a clearer error message when `length(condition) == 1`.
+
 # hutils 1.1.1
 * Change of stable test:
     - The only changes were so that the test proceeds when a required package is present. The tests would never have succeeded in cases where they now skip. Consequently the major version number has not been incremented.